Refactor OverlayRenderer
