--- warframe-relice-price/OverlayUI/hudUiElements/RelicPrices.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OverlayUI.hudUiElements;
-internal class RelicPrices
-{
-    public RelicPrices() 
-    { 
-
-    }
-}
-
