+++ warframe-relice-price/warframe-relice-price.csproj
@@ -19,8 +19,4 @@
     <PackageReference Include="Tesseract.Drawing" Version="5.2.0" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="OverlayUI\MenuUiElements\" />
-  </ItemGroup>
-
 </Project>
